+++ NINA/View/Equipment/Camera/CameraTemplateSelector.cs
@@ -20,15 +20,12 @@ namespace NINA.View.Equipment {
 
     internal class CameraTemplateSelector : DataTemplateSelector {
         public DataTemplate Default { get; set; }
-        public DataTemplate Touptek { get; set; }
         public DataTemplate QhyCcd { get; set; }
         public DataTemplate Altair { get; set; }
 
         public override DataTemplate SelectTemplate(object item, DependencyObject container) {
-            if (item is AltairCamera) {
+            if (item is ToupTekAlikeCamera) {
                 return Altair;
-            } else if (item is OmegonCamera || item is ToupTekCamera) {
-                return Touptek;
             } else if (item is QHYCamera) {
                 return QhyCcd;
             } else {